pyproject.toml
[project]
name = "support-triage-openenv"
version = "1.0.0"
description = "OpenEnv-compliant environment for automated customer support ticket routing"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "pytest",
    "pyyaml",
    "httpx",
]

[project.scripts]
server = "app:start_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["my_env", "app.py"]
